Value List update
I'm back on this feature, sorry for the delay. This is working with these modifications without making a double call to the DB, if you can review it that would be great.
I can not manage to get your tests up and running so I didn't modify your the LayoutTest. I can still provide you test access to an FM server if you want to.